requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.3.0
email-validator==2.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
firebase-admin==6.2.0
openai==0.28.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pytest==7.4.2
httpx==0.24.1
python-multipart==0.0.6
tenacity==8.2.3
numpy==1.25.2
requests==2.31.0